Remote XPubKey derivation + Ledger support
Couple of big features here:
- This depends on a fix for a bug currently existing on %zuse's 
+decompress-point arm for the secp crypto algorithm.
- A proposed fix is intended to be proposed as a pull request that 
allows to derive address from an extended (public) pubkey.
- This allows us to request a new key from a ship that we want to send 
BTC to, get it back on the frontend and sign a transaction with it.
- Another major features, still in the works, is to offer the option to 
derive the initial extended pubkey that's stored on urbit, via a Ledger 
Hardware Wallet. This is currently already implemented but needs to be 
further refined and integrated with the UI.
- Next step will be to convert BCoin's wallet into a "watch-only-wallet" 
and use the Ledger to sign the transactions.